requirements.txt
pyyaml
pandas
snowflake-snowpark-python
boto3
requests
numpy
python-dotenv